Code refactoring and prototype of pagination of content.
