+++ .idea/misc.xml
@@ -4,6 +4,7 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/Users/Mikolaj/AndroidStudioProjects/EpubMiniApp/app/src/main/res/layout/activity_main.xml" value="0.75" />
+        <entry key="..\:/Users/Mikolaj/AndroidStudioProjects/EpubMiniApp/app/src/main/res/layout/main_activity.xml" value="0.49150485436893204" />
         <entry key="..\:/Users/Mikolaj/AndroidStudioProjects/EpubMiniApp/app/src/main/res/layout/start_menu.xml" value="0.42109375" />
       </map>
     </option>
