+++ app/src/main/AndroidManifest.xml
@@ -18,6 +18,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".BookView"></activity>
     </application>
 
 </manifest>
