+++ app/src/main/assets/webassets/readium-default.css
@@ -27,8 +27,8 @@
   --RS__compFontFamily: var(--RS__baseFontFamily);
   --RS__codeFontFamily: var(--RS__monospaceTf);
 
-  --RS__typeScale: 1.125; /* 1.067 | 1.125 | 1.2 | 1.25 | 1.333 | 1.414 | 1.5 | 1.618 */
-  --RS__baseFontSize: 100%;
+  --RS__typeScale: 1.5; /* 1.067 | 1.125 | 1.2 | 1.25 | 1.333 | 1.414 | 1.5 | 1.618 */
+  --RS__baseFontSize: 150%;
 
   --RS__flowSpacing: 1.5rem;
   --RS__paraSpacing: 0;
