+++ app/src/main/res/layout/book_viewer.xml
@@ -4,13 +4,17 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".BookView">
 
     <LinearLayout
         android:layout_width="fill_parent"
         android:layout_height="fill_parent"
         android:orientation="vertical"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent">
+
         <LinearLayout
             android:layout_width="fill_parent"
             android:layout_height="50dip"
@@ -18,13 +22,12 @@
 
 
             <Button
-                android:id="@+id/btnPrev"
+                android:id="@+id/prev_btn"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_weight="1"
-                android:text="Previous"
                 android:onClick="prevChapter"
-                />
+                android:text="Previous" />
 
             <TextView
                 android:id="@+id/chapterCounter"
@@ -37,15 +40,13 @@
                 android:text="1/3" />
 
             <Button
-                android:id="@+id/btnNext"
+                android:id="@+id/next_btn"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_weight="1"
-                android:text="NEXT "
                 android:onClick="nextChapter"
-                />
+                android:text="NEXT " />
         </LinearLayout>
-
 
 
         <ImageView
@@ -53,6 +54,7 @@
 
             android:layout_width="fill_parent"
             android:layout_height="300dip"
+            android:layout_marginBottom="10px"
             android:layout_weight="0.1"
             tools:srcCompat="@tools:sample/avatars" />
